chore(path-structure): move types and aux functions to lib folder, update all the imports
