+++ src/lib/auxiliarFunctions.ts
